Save subprocess programs recursively
This will now take care of any issues with subprocess not being saved or
the wrong subprocesses being saved. It will save the newest subprocess.

I also added a default title for cmus at it would be swallowed reliably
-- before the title would be 'Alacritty' which would make it be
swallowed by other templates. This also makes it easy to assign cmus to
a workspace with i3 because the user can match the window by title.