Use error code to detect failures when attempting to save programs
The problem is in the subprocess.check_output function in Python, it
could print text, making the script detect that and display the nagbar.
This change fixes that by using error codes to detect when it fails,
allowing output to still be printed without the bash script detecting a
failure.