Simplify restoring
This is a big commit that enhances how i3-restore restores the last
session.

Now, i3-restore will save the command to restore each program in a
script. To restore, all that has to be done is execute that command
inside the corresponding container. This also solves problems with
i3-msg exec failing with commas and internal quotes inside shell
commands.

This also makes restoring sessions faster and more reliable.