Remove file extension for the two executable files
This is another breaking change. In your i3 config, you will need to
remove the file ending '.sh' when referring to either the i3-restore or
i3-save file.

This change was made to follow Google's shell style guide, which I will
use to format the code in the next commit.