save: Only save windows that have valid PIDs
I ran into an issue when saving my session with KeePass open because
the KeePass window doesn't have a PID as it is a subprocess. Now the
default behavior is to skip saving that container. The layout for that
window will still get restored, but it will be removed when the
kill_empty_containers function is called.

Also, switch from xprop to xdotool when getting the window PID. First,
the command is syntactically easier to understand. Also, some people
might not have xprop and we already depend on xdotool, so we might as
well use it.