Improve subprocess restoring
I tested this script out with the Kitty terminal and it did not work
correctly. Therefore, I have made a few adjustments for this script to
work in both Alacritty and Kitty.

Mainly, subprocess commands are now stored in files. This makes
executing them more reliable because they act almost identically to
executing them in the command line.