Ignore saving a container when it is a zombie process
This exception was thrown when i3-restore was saving my session as I
closed a window. My guess is that the process wasn't fully cleaned up
yet (still a zombie) and therefore i3-restore attempted to save it. Now,
the container will not be saved since it was just closed.