Fix KeyError while saving containers without a class property
The window that was giving me this issue was a Zoom meeting. I'm unsure
why the class doesn't appear (it only happens in meetings joined with a
link). The class appears with xdotool and xprop (so probably with X11)
but not with i3-msg -t get_tree. This may just be a small bug in i3-msg
or a difference in how these values are retrieved in X11 versus i3-msg.

Either way, the fix is simple. A test was also added to catch any
regressions.