Improve unit tests by reducing mocking
My position changed on this. I used to want to isolate the functions
that were being tested as much as possible by mocking every dependent
function in the test who's functionality is not being tested.

However, I think it is good to use the real implementation as much as
possible and only mock library functions where possible. This will allow
the unit tests to be more versatile by not needing to be adjusted for
every code change. Additionally, it will allow the unit tests to catch
more bugs which may only show up during chained calls.

Instead, integration tests can remove the mocks for the library
functions to allow the system to be better tested in more realistic
situations.