Rename --interval flag to --save-interval
This name is confusing as it is used with the restore script (mentioned
in #18), so this rename should help clear up the confusion.

Also, I formatted the usage menus to better accomodate for the longer
flag in the restore script while also not having too many spaces for the
save script.