Add support for plugins that use custom save algorithms for containers
This commit adds:
- Support for 'enabled_plugins' in the config so users can specify which
  plugins they want to use for these custom (more fine-grained) saves
- Saving containers that match a specific window class

As only custom saving for Kitty terminals is planned right now, only the
Kitty saving plugin will be created. However, the code is structured in
a way to make it easily extensible to more plugins.

The only places that need to be modified to add support for more plugins
are:
- The plugin module that contains the custom save algorithm. This main
  function in the module needs to take in the Container to be saved and
  the Config
- The SUPPORTED_PLUGINS dictionary in i3_save.py to point to the correct
  plugin module to use