Add window_class and window_id as attributes and move i3_PATH to utils
These attributes are needed by the Kitty plugin (at least, window_id is
but window_class still makes the code a bit cleaner and may be needed
too). Additionally, i3_PATH needs to be accessed outside of i3_save.py
so it is also moved to the general utils.py file.