Add basic Kitty plugin for restoring terminal sessions
The plugin restores Kitty OS windows and tabs. Additionally, it is
compatible with all subprocesses without the need to change your
subprocess config when you want to enable/disable the plugin.

In a future commit, more advanced restoring will be implemented,
including split window terminals and most other layouts you can do in
Kitty. However, it doesn't look like Kitty has information about this in
the tree, so these windows won't be able to be restored exactly in the
same layout.