Optionally restore Kitty terminal scrollback
Using Kitty's RC features, we can also retrieve the text from a Kitty
window. This gives us the option to restore the terminal scrollback
after a restore.

The options for the new 'scrollback' config option in the Kitty plugin
config accepts three options which determines how much of the scrollback
to restore:
- 'none' means to not restore any terminal scrollback
- 'screen' means to restore only the scrollback that was visible
- 'all' means to restore all scrollback for the terminal