Restore all windows in a Kitty tab
This restore isn't perfect, as Kitty's JSON tree provided by `kitty @
ls` doesn't provide all of the necessary information to reconstruct the
windows in a tab to the exact layout they are in (hsplit, vsplit,
location in the tab, etc.).

Therefore, each window in the Kitty tab will be restored in the layout
Kitty determines (which may or may not be correct). All subprocesses and
scrollback for each window will be restored still.