Don't fail restoring if the previously focused window does not swallow
If the window with the _i3_restore_focus mark does not swallow any
programs, it will be killed. This makes the i3-msg focus command fail
which causes the restore script to exit before finishing the restore
process.

Now, focus before any empty containers are killed to ensure a window
with the _i3_restore_focus mark exists. This is better than just
ignoring the error (e.g. using || true) as it will still restore focus
to the workspace (but not necessarily the exact window).