Add more testing coverage for workspaces with underscores
An underscore is possibly a special character because the script denotes
a separation in the file name with an underscore. There would be an
issue if the display name has an underscore in it, but I don't think
this actually happens with displays and will only fix if someone
actually runs into this issue.

One test was actually broken if a workspace had an underscore in its
name (it was how it was sanitizing the name in the test, not an issue
with the actual code). Some workspaces in the tests have been renamed to
include underscores to ensure this case is covered.